rename cache module to redis_cache
